feat: prefetch languages by CLI and use URL as source of truth
