+++ packages/cli/package.json
@@ -30,6 +30,7 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@shopware-pwa/shopware-6-client": "0.1.0-alpha.9",
     "gluegun": "^4.3.1",
     "lodash": "^4.17.15",
     "request": "^2.88.2",
