+++ packages/default-theme/components/SwLanguageSwitcher.vue
@@ -1,103 +1,45 @@
 <template>
-  <!-- TODO: change this if after onMounted hook here is resolved -->
-  <div
-    class="sw-language-switcher"
-    v-if="activeLanguage && availableLanguages.length > 1"
-  >
+  <div class="sw-language-switcher">
     <SfSelect
-      v-model="activeLanguage"
+      :selected="activeLanguage"
+      @change="redirectToLocale"
       :size="availableLanguages.length"
       class="sw-language-switcher__select"
-      @click="loadAvailableLanguages"
     >
       <SfSelectOption
-        v-for="currencyItem in availableLanguages"
-        :key="currencyItem.id"
-        :value="currencyItem.id"
+        v-for="language in availableLanguages"
+        :key="language.code"
+        :value="language.code"
       >
-        {{ currencyItem.translationCode.name }}
+        {{ language.name }}
       </SfSelectOption>
     </SfSelect>
   </div>
 </template>
 <script>
 import { SfSelect, SfProductOption } from '@storefront-ui/vue'
-import { useSessionContext } from '@shopware-pwa/composables'
 import { computed, onMounted, ref } from '@vue/composition-api'
-import {
-  getAvailableLanguages,
-  setCurrentLanguage,
-} from '@shopware-pwa/shopware-6-client'
+import languagesMap from 'sw-plugins/languages'
 
 export default {
   name: 'SwLanguageSwitcher',
   components: {
     SfSelect,
   },
-  setup(context) {
-    const { sessionContext, refreshSessionContext } = useSessionContext()
-
-    const availableLanguages = ref([])
-
-    const loadAvailableLanguages = async () => {
-      availableLanguages.value = await getAvailableLanguages()
-    }
-
-    // TODO: loaded on mounted only untill fixed issue: https://github.com/DivanteLtd/storefront-ui/issues/1097
-    onMounted(async () => {
-      await loadAvailableLanguages()
-    })
-
-    const activeLanguage = computed({
-      get: () =>
-        sessionContext.value &&
-        sessionContext.value.salesChannel &&
-        sessionContext.value.salesChannel.languageId,
-      // set: async (id) => await setCurrency({ id }),
-      set: async (id) => {
-        await setCurrentLanguage(id)
-        await refreshSessionContext()
-      },
-    })
-    return {
-      availableLanguages,
-      activeLanguage,
-      loadAvailableLanguages,
-      sessionContext,
-    }
-  },
-  watch: {
-    activeLanguage: {
-      immediate: true,
-      handler: function (val) {
-        if (!val || !this.availableLanguages.length) return
-        const code = this.availableLanguages.find((lang) => lang.id === val)
-        // console.error('ACTIVE LANGUAGE CHANGED!', code)
-        const xx = code && code.translationCode && code.translationCode.code
-        this.redirectToLocale(xx)
-      },
+  computed: {
+    availableLanguages() {
+      return Object.values(languagesMap)
     },
-    availableLanguages(languages) {
-      if (languages.length) {
-        // console.error('LANGUAGES', languages.length)
-        const curr = languages.find((lang) => lang.id === this.activeLanguage)
-        // console.error('CURRENT', curr)
-        const cirrCode = curr.translationCode && curr.translationCode.code
-        if (cirrCode !== this.$i18n.locale) {
-          // console.error('SHOULD SWITCH LOCALE', cirrCode)
-          this.redirectToLocale(cirrCode)
-        }
-      }
+    activeLanguage() {
+      return this.$i18n.locale
     },
   },
   methods: {
     redirectToLocale(localeCode) {
-      // console.error('Redirecting to locale', localeCode)
+      if (localeCode === this.$i18n.locale) return
       if (localeCode === this.$i18n.fallbackLocale) {
-        // console.error('STANDARD LANGUAGE!')
         this.$router.push(this.$route.fullPath.replace(/^\/[^\/]+/, ''))
       } else {
-        // console.error('CUSTOM LANGUAGE')
         this.$router.push(`/${localeCode}${this.$route.fullPath}`)
       }
     },
