+++ packages/default-theme/pages/_lang/account/addresses/add/_id.vue
@@ -1,0 +1,21 @@
+<template>
+  <div class="addresses-add">
+    <SwAddress />
+  </div>
+</template>
+
+<script>
+import SwAddress from '@shopware-pwa/default-theme/components/forms/SwAddress.vue'
+
+export default {
+  components: { SwAddress },
+  data() {
+    return {}
+  },
+  mounted() {
+    console.log(this.$router.params.id)
+  }
+}
+</script>
+
+<style lang="scss" scoped></style>
