+++ packages/shopware-6-client/src/services/contextService.ts
@@ -98,9 +98,7 @@ export async function setCurrentCurrency(
  * @throws ClientApiError
  * @alpha
  */
-export async function getAvailableLanguages(): Promise<
-  SearchResult<Language[]>
-> {
+export async function getAvailableLanguages(): Promise<Language[]> {
   const resp = await apiService.get(getContextLanguageEndpoint());
 
   return resp.data;
